+++ docs/components/guide/GuideBaseBadges.vue
@@ -153,7 +153,8 @@ export default {
             },
             details: {
               description: "Set the badge size.",
-              values: '"mini" | "small" | "default" | "medium" | "large"'
+              values:
+                '"micro" | "mini" | "small" | "default" | "medium" | "large"'
             }
           }
         ]